Introduce AOSPStarboardBridge base class for AOSP compatibility, restoring StarboardBridge name for Cobalt TV application.
