+++ cobalt/android/BUILD.gn
@@ -25,6 +25,7 @@ java_cpp_features("java_features_srcjar") {
 
 generate_jni("jni_headers") {
   sources = [
+    "apk/app/src/main/java/dev/cobalt/coat/AOSPStarboardBridge.java",
     "apk/app/src/main/java/dev/cobalt/coat/AudioPermissionRequester.java",
     "apk/app/src/main/java/dev/cobalt/coat/CaptionSettings.java",
     "apk/app/src/main/java/dev/cobalt/coat/CobaltService.java",
@@ -32,7 +33,6 @@ generate_jni("jni_headers") {
     "apk/app/src/main/java/dev/cobalt/coat/CobaltTextToSpeechHelper.java",
     "apk/app/src/main/java/dev/cobalt/coat/PlatformError.java",
     "apk/app/src/main/java/dev/cobalt/coat/ResourceOverlay.java",
-    "apk/app/src/main/java/dev/cobalt/coat/StarboardBridge.java",
     "apk/app/src/main/java/dev/cobalt/coat/javabridge/HTMLMediaElementExtension.java",
     "apk/app/src/main/java/dev/cobalt/features/StarboardFeatureList.java",
     "apk/app/src/main/java/dev/cobalt/media/AudioOutputManager.java",
@@ -116,12 +116,12 @@ android_library("cobalt_main_java") {
   ]
 
   sources = [
+    "apk/app/src/main/java/dev/cobalt/coat/AOSPStarboardBridge.java",
     "apk/app/src/main/java/dev/cobalt/coat/AdvertisingId.java",
     "apk/app/src/main/java/dev/cobalt/coat/ArtworkDownloader.java",
     "apk/app/src/main/java/dev/cobalt/coat/ArtworkDownloaderDefault.java",
     "apk/app/src/main/java/dev/cobalt/coat/ArtworkLoader.java",
     "apk/app/src/main/java/dev/cobalt/coat/AudioPermissionRequester.java",
-    "apk/app/src/main/java/dev/cobalt/coat/BrowserStarboardBridge.java",
     "apk/app/src/main/java/dev/cobalt/coat/CaptionSettings.java",
     "apk/app/src/main/java/dev/cobalt/coat/CobaltA11yHelper.java",
     "apk/app/src/main/java/dev/cobalt/coat/CobaltActivity.java",
