+++ cobalt/android/apk/app/src/app/java/dev/cobalt/app/CobaltApplication.java
@@ -16,7 +16,7 @@ package dev.cobalt.app;
 
 import android.app.Application;
 import android.content.Context;
-import dev.cobalt.coat.BrowserStarboardBridge;
+import dev.cobalt.coat.StarboardBridge;
 import org.chromium.base.ApplicationStatus;
 import org.chromium.base.ContextUtils;
 import org.chromium.base.PathUtils;
@@ -25,18 +25,18 @@ import org.chromium.base.library_loader.LibraryProcessType;
 
 /** Android Application hosting the Starboard application. */
 public class CobaltApplication extends Application
-    implements BrowserStarboardBridge.HostApplication {
+    implements StarboardBridge.HostApplication {
   private static final String PRIVATE_DATA_DIRECTORY_SUFFIX = "content_shell";
 
-  BrowserStarboardBridge mStarboardBridge;
+  StarboardBridge mStarboardBridge;
 
   @Override
-  public void setStarboardBridge(BrowserStarboardBridge starboardBridge) {
+  public void setStarboardBridge(StarboardBridge starboardBridge) {
     mStarboardBridge = starboardBridge;
   }
 
   @Override
-  public BrowserStarboardBridge getStarboardBridge() {
+  public StarboardBridge getStarboardBridge() {
     return mStarboardBridge;
   }
