+++ cobalt/android/apk/app/src/app/java/dev/cobalt/app/MainActivity.java
@@ -18,7 +18,7 @@ import android.app.Activity;
 import android.app.Service;
 import android.os.Bundle;
 import dev.cobalt.coat.ArtworkDownloaderDefault;
-import dev.cobalt.coat.BrowserStarboardBridge;
+import dev.cobalt.coat.StarboardBridge;
 import dev.cobalt.coat.CobaltActivity;
 import dev.cobalt.coat.CobaltService;
 import dev.cobalt.libraries.services.clientloginfo.ClientLogInfoModule;
@@ -33,11 +33,11 @@ import dev.cobalt.util.Holder;
 public class MainActivity extends CobaltActivity {
 
   @Override
-  protected BrowserStarboardBridge createStarboardBridge(String[] args, String startDeepLink) {
+  protected StarboardBridge createStarboardBridge(String[] args, String startDeepLink) {
     Holder<Activity> activityHolder = new Holder<>();
     Holder<Service> serviceHolder = new Holder<>();
-    BrowserStarboardBridge bridge =
-        new BrowserStarboardBridge(
+    StarboardBridge bridge =
+        new StarboardBridge(
             getApplicationContext(),
             activityHolder,
             serviceHolder,
